pyproject.toml
[project]
name = "z-image-turbo-demo"
version = "0.1.0"
description = "Z-Image Turbo GGUF 文生图 Demo for Apple Silicon Mac"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "pillow>=10.0.0",
]

[project.scripts]
generate = "generate:main"
server = "server.main:main"

[tool.setuptools]
py-modules = ["generate"]

[tool.uv]
python-downloads = "automatic"
